backend_site/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.4
attrs==21.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.5.0
distro-info==1.0
Django==3.2.4
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
django-webpack-loader==1.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
docopt==0.6.2
drf-flex-fields==0.9.1
duplicity==0.8.17
ecdsa==0.17.0
fasteners==0.14.1
feedparser==6.0.8
filelock==3.0.12
future==0.18.2
gyp==0.1
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.6.0
Jinja2==3.0.1
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mock==4.0.3
monotonic==1.5
netifaces==0.10.9
nose==1.3.7
npm==0.1.1
oauthlib==3.1.0
olefile==0.46
optional-django==0.1.0
packaging==21.0
paramiko==2.7.2
pbr==5.6.0
pexpect==4.8.0
Pillow==8.1.2
pipreqs==0.4.10
pluggy==0.13.1
protobuf==3.12.4
py==1.10.0
pyasn1==0.4.8
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.1.7+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.39
python-jose==3.3.0
python-termstyle==0.1.10
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
regex==2021.4.4
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
SecretStorage==3.3.1
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
sniffer==0.4.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
stevedore==3.3.0
systemd-python==234
testresources==2.0.1
toml==0.10.2
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.2
usb-creator==0.3.7
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.5
xdg==5
xkit==0.0.0
yarg==0.1.9
